update_2019/08/20

+++ Templates/数学/矩阵快速幂.cpp
@@ -1,6 +1,6 @@
 struct Matrix{
     int n;
-    LL mat[151][151];
+    long long mat[151][151];
     Matrix(){}
     Matrix(int _n) {
         n = _n;
